Kind-of solution
Problem is that now if it's throttling requests and you make an edit and wait, it's going to add a request to the queue with the latest value.

If you make another edit before that value is sent, it will append _another_ request on the queue.

Then it will send one request with the old value, then the other with the latest one. And the response of the old value might mess up the state a bit